Added a sub section in the troubleshooting.
I was trying to reproduce the workflow to understand how it can be done from scratch.
I did cargo build first and then flutter_rust_bridge_codegen. Of course I got the error "undefined symbol: store_dart_post_cobject". 
Finally I realized that the order is the other way around. The codegen should be done first and then cargo build (or make) so that the rust compiler can find the generated rust file and includes it to be a part of the end result.